+++ src/command_modules/azure-cli-network/azure/cli/command_modules/network/_help.py
@@ -3888,11 +3888,11 @@ helps['network watcher run-configuration-diagnostic'] = """
     examples:
         - name: Run configuration diagnostic on a VM with a single query.
           text: |
-            az network watcher run-configuration-diagnostic --target {VM_ID}
+            az network watcher run-configuration-diagnostic --resource {VM_ID}
                --direction Inbound --protocol TCP --source 12.11.12.14 --destination 10.1.1.4 --port 12100
         - name: Run configuration diagnostic on a VM with multiple queries.
           text: |
-            az network watcher run-configuration-diagnostic --target {VM_ID}
+            az network watcher run-configuration-diagnostic --resource {VM_ID}
                 --queries '[
                 {
                     "direction": "Inbound", "protocol": "TCP", "source": "12.11.12.14",
@@ -4018,7 +4018,7 @@ helps['network watcher packet-capture create'] = """
         - name: --vm
           short-summary: Name or ID of the VM to target.
         - name: --filters
-          short-summary: JSON encoded list of packet filters. Use `@<file path>` to load from file.
+          short-summary: JSON encoded list of packet filters. Use `@{path}` to load from file.
     examples:
         - name: Create a packet capture session on a VM.
           text: az network watcher packet-capture create -g MyResourceGroup -n MyPacketCaptureName --vm MyVm --storage-account MyStorageAccount